Move parcel calculations to separate file
